Fix several bugs in RangeSource operator.
Before this fix, the RangeSource operator was (non-sensically) compiled
into a FilterOperator. With this fixed, then plan would still not
compile because the constants from the range could not be converted to
tuples, which is what the execution operator expected. With this fix, it
expects the type of the first value of its tuple type, which could be a
long or a double. The handling of different types is still a bit sloppy:
longs are converted to doubles and back, which is lossfull for big
values.