Fix TestReduceByKey.test_grouped.
The previous test used something like `t1[0] + t2[0]`, which was meant
to add to numbers, but, since the input tuples are nested, this code
really concatenates tuples. The previous commit made that visible. The
new version first flattens the tuple such that `+` means addition.
Furthermore, the reference result is now computed using python
constructs (instead of consisting of a wrong, hard-coded result).