Dedicated docker image with automated build and Python packages.
This commit uses a docker image for CI based on an automated build. This
makes it transparent what the image consists of and allows others to
change the docker image more easily. Also, the new image also installs
the current Python dependencies, which took about 2 minutes on every CI
test previously.