Code reorganization to allow generalizing parser tests.
The test of the parser just tested whether parsing round-trips: the
output of the parser had to be the same as its input. Now, the test code
is more general: it tests whether the optimizer tool produces a given
output. The parser tests are now a special case, namely running the
optimizer with all optimizations disabled. The output is now given as
files, which are exact copies of the input files. Finally, the test now
includes a tool that allows to reproduce the output files to what the
optimizer currently produces to set that is the new reference.