Fix: Rewrite move logic of SimplePredicateMoveAround.
The new logic is more conservative in the sense that it doesn't touch
filters that would be moved to the place where they already are. This
only changes their ID and formatting of the LLVM IR code, but produces
otherwise isomorphic DAGs. This commit applies these (formatting)
changes to the optimizer test cases.