Fix printDAG: allow parallel edges between two nodes.
There could be two parallel edges between two nodes, e.g., a self-join.
This is supported by cgraph, but adding requires to give different names
to the different parallel edges. If the name is the same, agedge has
lookup semantics and does not add a second edge. This commit thus
computes a name based on input and output ports for each edge.