Implement final optimizer pass that canonicalizes the DAG.
This means that is assigns operator IDs that are independent of their
creation order (which could depend on machine details such as the hash
function of an unordered_map) and then orders the vertices and edges
physically according to these IDs. This should bring the graph in a
canonical form that is independent of how it was produced.

This is necessary for testing, where "the same DAG" could have different
forms on different machines, but will also be useful for caching plans.