Reduce partitioning-fanout from 256 to 16.
This is a temporary solution needed to make CI pass again. The root
cause is a memory leak in the materialization operator, whose fix is now
tracked in #48. Since the result of each partition is materialized, the
partitioning fan-out influences the amount of memory leaked. On the CI
machines, the leaks where so high that the machine ran out of memory,
which in turn prevented `fork`s and thus running external commands. With
a lower fan-out, the leak is small enough to let the tests pass.

This fixes #44.