Handle LLVM IR produced by NUMBA correctly.
Until now, we extraced the function we thought was the only part of the
produced IR we needed and concatenated these into a single file during
codegen. However, we lost attributes of the function that way (which we
had to add back with a special AddAlwaysInline optimizer pass). Also,
this stripped the declaration of external functions that NUMBA may add,
which considerably limits the set of functions we can support. Now, we
keep the entire output of NUMBA and use llvm-link to merge them.