Separate flattening from JSON encoding of tuples in Python.
Currently, the JSON encoder of types in the Python level transforms the
types in order to make them flat. This makes it impossible to encode (1)
nested types and (2) types that aren't in a tuple at all, e.g., a plain
`bool`. The latter is needed soon in order to encode the return type of
UDFs, which is `bool` in case of filters.