Change configuration of Boost graph structure.
We thought that we were using std::list for vertex storage (as this is
recommended when doing frequent updates). However, the template
parameters are actually in a different order, so std::list was used for
edges. We cannot use std::list for vertex because they do not support
"internal properties", which we need for storing DAGOperators with
vertexes.