Fix inlining test case.
For a while now (!100), the inlining test was broken, as it assumed the
generated library in a "known" place, but they are now in a different
place, so the current mechanism doesn't find them anymore (see #63).
This commit makes `opt` return the generated library names, which the
inlining test picks up to find the files.

This solves the basic point of #63.