Make debug and opt version of generated code co-exist.
If the same query is run twice, once in debug mode and once in opt mode,
the second will overwrite the shared library of the first, which may
lead to crashes if some part of the library needs to be reread from
disk. This commit leaves the two version as `.opt.so` and `.dbg.so` such
that the two can co-exist.