Fix memory leak in runtime::Free.
Turns out that our implementation of Free was bogus and never freed
anything. The problem was that the data to be free was shipped in a
runtime::Tuple, but we expected a runtime::Array directly. This was
implemented with a visitor, which doesn't have a base case and thus
silently ignores unknown types, in this case tuples.