Chose CXX11 ABI explicitly in backend and JIT compilation.
This makes sure that the jitted code uses the same version of the ABI of
libstdc++, which might not be the case if it is compiled on a different
machine than the backend. This in turn can happen if the drivers ships
the jitted code to a different machine (which does happen in the
lambdas).