Make polling strategy in S3 exchange more aggressive.
This is a performance regression introduced while refactoring the code
from the initial implementation (used in the Modularis paper) before
merging it to master. The regression introduced a 50% longer running
time in some queries. As a down side, the more aggressive polling has
probably a higher monetary cost (which I haven't studied).