Fix two-level partitioning: use two different hash seeds.
This is a regression introduced while refactoring the initial
implementation (used in the Modularis paper) for merging to master. It
is relevant for performance since it balances the work badly in any
two-level exchange. In such a case, the data should be partitioned
twice, using a different hash function every time. The regression used
the same hash function in both levels.