Delete temporary files at the beginning and end of the Lambda handler.
This should prevent /tmp to fill up and render the serverless workers
unable to work. Deleting the files at the beginning of the handler might
be necessary in case the previous execution could not do it at its end.
Deleting them at the end, however, should be the common case, as it is
not on the critical path to running queries.